+++ gframe/tag_duel.cpp
@@ -496,7 +496,7 @@ void TagDuel::Surrender(DuelPlayer* dp) {
 int TagDuel::Analyze(char* msgbuffer, unsigned int len) {
 	char* offset, *pbufw, *pbuf = msgbuffer;
 	int player, count, type;
-	while (pbuf - msgbuffer < len) {
+	while (pbuf - msgbuffer < (int)len) {
 		offset = pbuf;
 		unsigned char engType = BufferIO::ReadUInt8(pbuf);
 		switch (engType) {
